Change how we determine branches
This requires that remote branches follow a naming scheme of
<release>/master and <release>/topic where release is like "f13" or
"epel6".  If we don't find a f<something, epel<something>, or
olpc<something>, we assume that it is either master, or a branch of
master for rawhide, and treat it like rawhide.  To determine rawhide
target we look at all the f branches, find the latest one and add 1 to
the number.  Seems to work and will keep working until Fedora 99.