fedpkg: Try not to add redundant gitignore entries
If there is a entry which matches the filename we're about to add
already in the gitignore file, we don't need to bother adding another
entry.

This adds a match() method to GitIgnore which uses fnmatch.  Under the
hood, this is what git uses.  It doesn't catch some of the special cases
git has, but it should simply fail to match in those cases and will add
a potentially redundant entry -- no different than previously.