Update the documentation for a lot of commands
This feeds both command --help and the man page.  Formatting is a little
odd.  The help = on the parser itself cannot span multiple lines with \.
It'll look fine in --help but not in the man page.  Multiple strings
need to be used instead.  The description argument can span multiple
lines with \ just fine though.